🚚 move: 이벤트 개설 페이지 컴포넌트 폴더구조 변경
