+++ src/components/openEvent/atoms/Header/Header.style.ts
