+++ src/components/openEvent/atoms/Label/Label.style.ts
