+++ src/components/openEvent/atoms/OpenEvent.style.ts
