+++ src/components/openEvent/atoms/index.ts
@@ -2,6 +2,7 @@ export { default as DateTimePicker } from './DateTimePicker/DateTimePicker';
 export { default as FindButton } from './FindButton/FindButton';
 export { default as Header } from './Header/Header';
 export { default as HelpText } from './HelpText/HelpText';
+export { default as ImageButton } from './ImageButton/ImageButton';
 export { default as Input } from './Input/Input';
 export { default as Label } from './Label/Label';
 export { default as Tag } from './Tag/Tag';
