+++ src/components/openEvent/forms/Address/Address.tsx
@@ -5,7 +5,7 @@ import { OpenEvent } from 'components/openEvent';
 import useNavigator from 'hooks/navigator/useNavigator';
 import { useOpenEventStore } from 'stores/OpenEventStore';
 import StatusType from 'constants/status';
-import { HelpText } from 'components/openEvent/OpenEvent';
+import { HelpText } from 'components/openEvent/atoms';
 import addressStyles from './Address.style';
 
 const Address = () => {
