+++ src/components/openEvent/forms/Period/Period.tsx
@@ -3,7 +3,7 @@ import { ScrollView, View } from 'react-native';
 import { OpenEvent } from 'components/openEvent';
 import { useOpenEventStore } from 'stores/OpenEventStore';
 import { serverDateFormatter } from 'utils/date';
-import { HelpText } from 'components/openEvent/OpenEvent';
+import { HelpText } from 'components/openEvent/atoms';
 import StatusType from 'constants/status';
 import periodStyles from './Period.style';
