+++ src/components/openEvent/index.ts
@@ -1,2 +1,2 @@
-export * as OpenEvent from './OpenEvent';
+export * as OpenEvent from './atoms';
 export * as OpenEventForm from './forms';
